requirements.txt
bcrypt==4.0.1
click==8.1.3
Flask==2.2.2
Flask-MySQLdb==1.0.1
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
PyJWT==2.6.0
Werkzeug==2.2.2